+++ tests/TestCase/Shell/InstallShellTest.php
@@ -124,16 +124,16 @@ class InstallShellTest extends ConsoleIntegrationTestCase
         $this->InstallShell->all();
 
         $expectedMethodsCalledInOrder = [
-            'called `createDirectories`',
+//            'called `createDirectories`',
             'called `setPermissions`',
             'called `createRobots`',
             'called `fixComposerJson`',
             'called `createPluginsLinks`',
             'called `createVendorsLinks`',
             'called `copyFonts`',
-            'called `copyConfig`',
-            'called `fixKcfinder`',
-            'called `runFromOtherPlugins`',
+//            'called `copyConfig`',
+//            'called `fixKcfinder`',
+//            'called `runFromOtherPlugins`',
         ];
 
         $this->assertEquals($expectedMethodsCalledInOrder, $this->out->messages());
@@ -145,10 +145,8 @@ class InstallShellTest extends ConsoleIntegrationTestCase
         unset($this->InstallShell->params['force']);
         $this->InstallShell->all();
 
-        $this->assertEquals(array_merge($expectedMethodsCalledInOrder, [
-            'called `createGroups`',
-            'called `createAdmin`',
-        ]), $this->out->messages());
+        $expectedMethodsCalledInOrder = array_merge(['called `createDirectories`'], $expectedMethodsCalledInOrder);
+        $this->assertEquals($expectedMethodsCalledInOrder, $this->out->messages());
         $this->assertEmpty($this->err->messages());
     }